Reorganize select anchor services and finish replace_side
